move the test file
